Improve implementation hint in InefficientEmptyStringCheck
The alternative implementation mentions iterating over the string and
checking for whitespace which is pretty clear, however, showing the
implementation doesn't hurt, nor does the hint for Apache
commons-lang's StringUtils.isBlank which does exactly that.

The if statement in the demonstration of inefficient code can lead to
confusion because users _might_ think that the check is only
inefficient inside an if statement although it's clear from the
description that it's not.