Fix ConcurrentModificationException on designer
It was triggered when restoring multiple properties,
since the items where added to a collection that was
being iterated.

Also fix a small bug when creating a property named
TODO -> the next property was allowed to be called
TODO if the name wasn't changed, since the name field
wasn't changed, so didn't trigger reevaluation of the
validators.